Update libswscale;
git-svn-id: https://mpc-hc.svn.sourceforge.net/svnroot/mpc-hc/trunk@3887 10f7b99b-c216-0410-bff0-8a66a9350fd8

+++ src/filters/transform/MPCVideoDec/ffmpeg/libswscale/utils.c
@@ -532,7 +532,7 @@ fail:
 }
 
 #if HAVE_MMX2
-static int initMMX2HScaler(int dstW, int xInc, uint8_t *filterCode, int16_t *filter, int16_t *filterPos, int numSplits)
+static int initMMX2HScaler(int dstW, int xInc, uint8_t *filterCode, int16_t *filter, int32_t *filterPos, int numSplits)
 {
     uint8_t *fragmentA;
     x86_reg imm8OfPShufW1A;
